requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
APScheduler==3.10.4
SQLAlchemy==2.0.32
psycopg[binary]==3.1.19
requests==2.32.3
pytz==2024.1
